services/agent_adk/requirements.txt
# Core dependencies extracted from pyproject.toml
requests>=2.31.0
google-adk>=1.11.0
pydantic>=2.11.3
python-dotenv>=1.1.0
google-cloud-aiplatform[adk,agent_engines]>=1.128.0
absl-py>=2.1.0
cloudpickle>=3.0.0
resend>=2.0.0
reportlab>=4.0.0
aiohttp>=3.9.0

# Development dependencies
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0